server/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
transformers
torch
#scikit-learn
#pydantic
python-dotenv
